requirements.txt
pyrogram>=2.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
motor>=2.5.0